Initial frontend layout (#12)
* Finally complies with <Auth> tags.

Needs to be cleaned in Program.cs'es from unnecessary lines. look see in usings also in _imports and etc to clean from uneccesary. IDK why on server side after pre-rendering everything 'not found'?

* /tickets page and /tickets/create frontend sample generated succefsully

* ready for calling the api. structure of folders still needs improvements. It seems like services are in client project while need to be on server's one.

Commit before changing the backend for cors politics

* The front with back are now friends.

However, I've changed the back, the Program.cs. And now the conflicts are very may be may be